chapter_5/pyproject.toml
[project]
name = "chapter-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "ipython>=9.8.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
